+++ src/api/v2/position.rs
@@ -1,4 +1,4 @@
-// Copyright (C) 2019-2020 Daniel Mueller <deso@posteo.net>
+// Copyright (C) 2019-2021 Daniel Mueller <deso@posteo.net>
 // SPDX-License-Identifier: GPL-3.0-or-later
 
 use std::ops::Not;
@@ -224,7 +224,7 @@ mod tests {
     let api_info = ApiInfo::from_env().unwrap();
     let client = Client::new(api_info);
     let symbol = asset::Symbol::Sym("SPY".to_string());
-    let result = client.issue::<Get>(symbol).await;
+    let result = client.issue::<Get>(&symbol).await;
 
     // We don't know whether there is an open position and we can't
     // simply create one as the market may be closed. So really the best