+++ src/api/v2/positions.rs
@@ -1,4 +1,4 @@
-// Copyright (C) 2019-2020 Daniel Mueller <deso@posteo.net>
+// Copyright (C) 2019-2021 Daniel Mueller <deso@posteo.net>
 // SPDX-License-Identifier: GPL-3.0-or-later
 
 use crate::api::v2::position::Position;
@@ -42,6 +42,6 @@ mod tests {
     // reporting any errors.
     let api_info = ApiInfo::from_env().unwrap();
     let client = Client::new(api_info);
-    let _ = client.issue::<Get>(()).await.unwrap();
+    let _ = client.issue::<Get>(&()).await.unwrap();
   }
 }