+++ CHANGELOG.md
@@ -1,6 +1,7 @@
 Unreleased
 ----------
 - Added `ApiInfo::from_parts` constructor
+- Adjusted `Client::issue` to accept request input via reference
 - Introduced `ConversionError` type
   - Replaced `unwrap`s with proper error variants
 - Updated `ActivityType` enum to be in sync with upstream variants